requirements.txt
langchain
torch
accelerate
sentence_transformers
streamlit_chat
streamlit
faiss-cpu
tiktoken
ctransformers
huggingface-hub
pypdf
python-dotenv
replicate
docx2txt
openai
requests
smtplib
time
nltk
sklearn
langchain
os
dotenv
tempfile


